Change Log: - Additional Controller/Services/Repository methods implemented: - Get Products by name; - Get Products by id; - Get Users by email.
Additional Changes:
- Updated methods (Controller) to return ResponseEntity's (JSON formatting).
- Added Address model class;
- Commented out latitude & longitude columns in UserAddresses table, as not needed unless delivery feature to be implemented & can easily revert change.

+++ src/main/java/au/edu/rmit/sept/SuperPrice/repository/ProductRepository.java
@@ -19,9 +19,9 @@ public interface ProductRepository extends PagingAndSortingRepository<Product, L
     // Find all Products
     List<Product> findAllBy();
 
-    // TODO: Find Products by product_name
+    // TODO: Find Products by product_name -> Currently getting the product id only
     @Query(value = "SELECT p.product_id FROM Product p WHERE LOWER(p.product_name) = LOWER(:product_name)")
-    int retrieveProductIdByProductName(@Param("product_name") String product_name);
+    Long retrieveProductIdByProductName(@Param("product_name") String product_name);
 
 //    @Query(value = "SELECT p.product_name, p.product_id, p.product_category FROM Product p WHERE LOWER(p.product_name) = LOWER(:product_name)")
 //    Product findByName(@Param("product_name") String product_name);
